Vendor diff and cmp libraries.
Lots of hard work has been done on these already.  go-wish just wants
to provide some additional opinionation about usability in tests
which make these libraries easier to reach for.

I'm vendoring them because:

A) Holistically, I simply want go-wish to be a single-repo dependency.
It's just plain easier to manage this way.

B) Some of the features I intend to use in these libraries are
marked as "don't rely on output", and I nonetheless intend to do so!
So, supporting ranges of versions of the libraries is off the menu.

Licenses of the libraries are retained in their respective directories.

Hopefully, these will be vendoring without significant patching.
No promises, though.  Consistency of output in go-wish overall will
be the overriding concern if push comes to shove.

Signed-off-by: Eric Myhre <hash@exultant.us>