Introduce ShouldEqual, and a series of examples.
ShouldEqual almost entirely defers the hard work to the sensible
behaviors already implemented in go-cmp, except to use multi-line
diffs for strings, which is often more useful.

The output is not yet polished or consistent.  We'll be building
up a battery of indentation mungers for that, for better or for
worse.

Signed-off-by: Eric Myhre <hash@exultant.us>