More t.Log interaction fixtures.
Doing some checks with types that go through go-cmp rather than the
string diff shortcircuit.  Want to make sure these have equivalent
spacing.  They do.

I also suddenly noticed that we're effectively getting double indent
with stdlib testing.T.Log, because *we* indent it, and then stdlib
does it *again* for all the non-first lines.  You don't see this in
our `Example*` tests, because the fakeT.Log function isn't a faithful
reproduction of stdlib's -- something we should probably fix, for
the sake of more realistic examples and docs if nothing else.
I'm undecided if the double indent is a problem worth addressing, or
an accidental feature.

Signed-off-by: Eric Myhre <hash@exultant.us>