'wishfix' file format unmarshaller implementation.
And many, many tests for the edge cases of parsing.

Note that tests of error messages are a little displeasing.  I'd prefer
to make those assert equality to another `fmt.Errorf`.  This would have
the happy effect of immediately Doing The Right Thing even if the value
is nil -- as it is, these tests will panick if the testy thing actually
*passes* unexpectedly.  Unfortunately, we can't do this.  go-cmp has
Opinions about unexported fields that will *also* panic here without
either panics, or making some Special Cases to our compare method which
is also a road I don't wanna go down.  Hrm.  Problems for Later.
So, these tests are less than perfect in their outcomes when they fail
(they'll panic if errors are unexpectedly nil).

Signed-off-by: Eric Myhre <hash@exultant.us>