Patch go-cmp to allow unexported comparisons.
I understand and to a large degree agree with the go-cmp library's
design decision to keep comparison of exported fields a feature which
one should be wary of.  And if I was using this library in any code
which actually *ships to production*, I'd indeed be incredibly ginger
with this power.  However, what we're writing here is a test library.
And that has a different set and degree of concerns.

Tl;dr I've got things to do, and comparing errors is top of that list.

As the last couple of commits (and the doc comments on `wish.Require`)
have commented, one of the most bog-standard things we need to do when
writing tests in Go is write some very basic checks on errors.
A huge fraction of the time, this involves comparing the unfortunately
unexported errors.errorString type.  And yet, I'm utterly loathe for
go-wish to make a special case of pre-configuring our use of go-cmp
to be unusually tolerant towards that type; we regress completely on
the portability to non-unsafe platforms ("purego") platforms by making
an use of this feature (so, in for a penny, in for a pound); go-cmp
doesn't currently export any way to configure it to allow *any*
(only specific types); and we don't particularly want to hoist the
responsibility of configuring such a bog standard thing as sensible
flags to make error comparing "Do What I Mean" all the way to a user
if we want to have a testing library with at all sensible verbosity.

So!  We have a use case; we have no good excuse for punting on it;
and we have no way to do what we need in the API we've been given.

Thus: We're making a tiny patch to go-cmp that does the thing.

And per the "in for a penny, in for a pound" logic stated above,
this does apply to all fields on all types.  I think ignoring fields
still has priority, though, if anyone ever needs that for some reason.

Interesting side note 1: the well known 'go-spew' package also uses
somewhat similar unexported/unsafe hacks.  They live in 'bypass.go'.

Interesting side note 2: though I was wary the implications of
the unexported field traversal being difficult to presume on "purego"
environments, I'm assured today by a gopherjs maintainer (who I'll not
name because I'm uncertain this was supposed to be an "on the record"
sort of thing) that even there, "it's just go", and it's somewhat
unusual to feel a strong need to run your tests on the far side.
So that makes me feel significantly less concerned about this kind of
portability restriction in a testing library, which is nice.

Signed-off-by: Eric Myhre <hash@exultant.us>