Update string diff report format.
Since go-cmp has changed its report syntax somewhat, and our syntax was
based on an attempt to be roughly congruous with the go-cmp syntax, so
too should ours now change.

The net indentation level is decreased.  Again, in the name of visual
consistency.

The diff indicators are adjusted!  (This required a new patch, this
time to the difflib library.)  They are now composed of a plus or minus
or space *followed by another space*: this is the same as the go-cmp
diff indicators.  (It also just plain looks a little less noisy, IMHO.)
This does technically make the diffs incompatable with `patch`;
but we've never been particularly concerned about direct interface
with such tools, and are not about to start fretting overly now.

Signed-off-by: Eric Myhre <hash@exultant.us>